# Faster padding - fuse fill, reduce overhead