--- a/nestedtensor/csrc/masking.cpp
+++ b/nestedtensor/csrc/masking.cpp
@@ -505,12 +505,13 @@ Tensor to_padded_tensor(Tensor nt, double padding) {
       Tensor offsets = batch_offsets_from_efficient_size(esize);
       std::vector<int64_t> new_size = padded_size_from_efficient_size(esize);
       at::cuda::CUDAStream defaultStream = at::cuda::getDefaultCUDAStream();
-      Tensor output;
-      if (padding == 0) {
-        output = at::zeros(IntArrayRef(new_size), nt_buffer.options());
-      } else {
-        output = nt_buffer.new_full(IntArrayRef(new_size), padding, nt_buffer.options());
-      }
+      // Tensor output;
+      Tensor output = at::empty(IntArrayRef(new_size), nt_buffer.options());
+      // if (padding == 0) {
+      //   output = at::zeros(IntArrayRef(new_size), nt_buffer.options());
+      // } else {
+      //   output = nt_buffer.new_full(IntArrayRef(new_size), padding, nt_buffer.options());
+      // }
       Tensor new_size_tensor = torch::tensor(new_size);
 
       int64_t input_dim = nt_sizes.size(1);
@@ -533,6 +534,7 @@ Tensor to_padded_tensor(Tensor nt, double padding) {
         nested_tensor::cuda::add_padding_kernelLauncher(
             nt_buffer.data_ptr<c10::Half>(),
             output.data_ptr<c10::Half>(),
+            (c10::Half)(padding),
             offsets.data_ptr<int>(),
             nt_sizes.data_ptr<int>(),
             input_dim,
@@ -545,6 +547,7 @@ Tensor to_padded_tensor(Tensor nt, double padding) {
         nested_tensor::cuda::add_padding_kernelLauncher(
             nt_buffer.data_ptr<float>(),
             output.data_ptr<float>(),
+            (float)(padding),
             offsets.data_ptr<int>(),
             nt_sizes.data_ptr<int>(),
             input_dim,