--- a/nestedtensor/version.py
+++ b/nestedtensor/version.py
@@ -1,5 +1,5 @@
-__version__ = '0.1.4+036a159'
-git_version = '036a15938527fbf84805e96c9679cf7e4d432b0c'
+__version__ = '0.1.4+54117b5'
+git_version = '54117b58363575602eab13131808959482a13511'
 from nestedtensor import _C
 if hasattr(_C, 'CUDA_VERSION'):
     cuda = _C.CUDA_VERSION